second commit

+++ src/main/java/com/minakov/entity/carShowroom/CarShowroom.java
@@ -5,19 +5,19 @@ import com.minakov.entity.Entity;
 public class CarShowroom extends Entity {
 
     private String street;
-    private short houseNumber;
+    private int houseNumber;
     private String city;
 
     public CarShowroom() {
     }
 
-    public CarShowroom(String street, short houseNumber, String city) {
+    public CarShowroom(String street, int houseNumber, String city) {
         this.street = street;
         this.houseNumber = houseNumber;
         this.city = city;
     }
 
-    public CarShowroom(int id, String street, short houseNumber, String city) {
+    public CarShowroom(int id, String street, int houseNumber, String city) {
         setId(id);
         this.street = street;
         this.houseNumber = houseNumber;
@@ -32,11 +32,11 @@ public class CarShowroom extends Entity {
         this.street = street;
     }
 
-    public short getHouseNumber() {
+    public int getHouseNumber() {
         return houseNumber;
     }
 
-    public void setHouseNumber(short houseNumber) {
+    public void setHouseNumber(int houseNumber) {
         this.houseNumber = houseNumber;
     }
